Create a specific user and group to execute the testsuite
The alternative implementation would have been to create the user during
image building to use the USER directive in the Dockerfile. The drawback
of this is duplication. We would have to apply this change across all
Dockerfiles and to every new Dockerfile as well.

docker-entrypoint now just takes care of setting up the user and
launches the actual testsuite as a user. The split is necessary to
prevent having to execute every command with su cgal_tester -c and
horrible escaping and environment interference.

Fixes #12